Add files via upload
This is it, the bread and butter of this project. This python routine uses the MeCab wrapper to analyze text. I also added a function at the end of the routine to insert the vocab words back into a postgres database. I'll upload some more accompanying files to the repository after this.