Add files via upload
Too many of the titles in my database where not found in The Movie Database so, I became a patreon to the Open Movie Database. Here, I use their API to get genres for titles and then add them to the postgres database. Since the OMDB does not have genre id codes, I will reverse engineer the many-to-many table to create a one to many table for genre ids.